I implemented an async middleware function which takes a query value (ex. Hello world) and send a message in a Slack channel. Now whenever a user POST to /messages with param === text, it will be send to a channel. This is a simple test so I can see if there is any troubles on passing values between the Client - Gateway - Service. Everything seems fine.
